requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
APScheduler==3.11.1
asyncpg==0.31.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
environs==14.5.0
Flask==3.1.2
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.8.0
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.1.0
multidict==6.7.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.2.1
redis==7.1.0
SQLAlchemy==2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
waitress==3.0.2
Werkzeug==3.1.4
WTForms==3.2.1
yarl==1.22.0
